Move color converter to parse method
